refactor: Class Contest 폴더 구조 변경
