+++ src/pages/Class/Contest/api/index.ts
@@ -1,4 +1,4 @@
-import { api } from '@/api';
+import { api, fileApi } from '@/api';
 
 const API_URL = `/class`;
 
@@ -6,20 +6,8 @@ const getContestList = (classId: string) => {
   return api.get(`${API_URL}/${classId}/contests`);
 };
 
-const getContestProblemList = (classId: string, contestId: string) => {
-  return api.get(`${API_URL}/${classId}/contests/${contestId}`);
-};
-
 const createContest = ({ classId, payload }: { classId: string; payload: ContestRequest }) => {
   return api.post(`${API_URL}/${classId}/contests/`, payload);
-};
-
-const editContestVisible = (classId: string, contestId: string) => {
-  return api.patch(`${API_URL}/${classId}/contests/${contestId}/check/`);
-};
-
-const deleteContest = (classId: string, contestId: string) => {
-  return api.delete(`${API_URL}/${classId}/contests/${contestId}/`);
 };
 
 const editContest = ({
@@ -34,11 +22,60 @@ const editContest = ({
   return api.patch(`${API_URL}/${classId}/contests/${contestId}/`, payload);
 };
 
+const editContestVisible = (classId: string, contestId: string) => {
+  return api.patch(`${API_URL}/${classId}/contests/${contestId}/check/`);
+};
+
+const deleteContest = (classId: string, contestId: string) => {
+  return api.delete(`${API_URL}/${classId}/contests/${contestId}/`);
+};
+
+const getContestProblemList = (classId: string, contestId: string) => {
+  return api.get(`${API_URL}/${classId}/contests/${contestId}`);
+};
+
+const getContestProblem = ({ classId, contestId, contestProblemId }: ContestProblemRequest) => {
+  return api.get(`${API_URL}/${classId}/contests/${contestId}/${contestProblemId}`);
+};
+
+/** FIXME: 페이지네이션, username */
+const getContestProblemSubmission = ({
+  classId,
+  contestId,
+  contestProblemId,
+}: ContestProblemRequest) => {
+  return api.get(`${API_URL}/${classId}/contests/${contestId}/${contestProblemId}/submissions`);
+};
+
+const createContestProblemSubmission = ({
+  classId,
+  contestId,
+  contestProblemId,
+  payload,
+}: ContestProblemRequest & { payload: FormData }) => {
+  return fileApi.post(
+    `${API_URL}/${classId}/contests/${contestId}/${contestProblemId}/submission/`,
+    payload
+  );
+};
+
+const createContestProblemSumbissionCheck = ({
+  classId,
+  contestId,
+  contestProblemId,
+}: ContestProblemRequest) => {
+  return api.patch(`${API_URL}/${classId}/contests/${contestId}/${contestProblemId}/check/`);
+};
+
 export {
   getContestList,
-  getContestProblemList,
   createContest,
+  editContest,
   editContestVisible,
   deleteContest,
-  editContest,
+  getContestProblemList,
+  getContestProblem,
+  getContestProblemSubmission,
+  createContestProblemSubmission,
+  createContestProblemSumbissionCheck,
 };
