+++ src/pages/Class/Contest/components/index.ts
@@ -1,3 +1,6 @@
 export * from './ContestFormModal';
 export * from './ContestEditModal';
 export * from './ContestProblemCreateModal';
+
+export * from './FileSubmissionForm';
+export * from './LeaderboardSubmissionForm';
