+++ src/pages/Class/Contest/hooks/query/index.ts
@@ -4,3 +4,8 @@ export * from './useCreateContestMutation';
 export * from './useDeleteContestMutation';
 export * from './useEditContestMutation';
 export * from './useEditContestVisibleMutation';
+
+export * from './useContestProblemQuery';
+export * from './useContestProblemSubmissionQuery';
+export * from './useCreateContestProblemSubmissionMutation';
+export * from './useCreateContestProblemSubmissionCheckMutation';
