+++ src/pages/Class/Contest/index.ts
@@ -1,3 +1,7 @@
 export * from './ClassContest';
 export * from './ClassContestProblemList';
 export * from './ClassEditContestList';
+
+export * from './ContestProblemDetail';
+export * from './ContestProblemLeaderBoard';
+export * from './ContestProblemSubmission';
