+++ src/pages/Class/Problem/components/index.ts
@@ -1,4 +1,2 @@
 export * from './Problem';
-export * from './FileSubmissionForm';
-export * from './LeaderboardSubmissionForm';
 export * from './ProblemForm';
