+++ src/pages/Class/Problem/hooks/query/index.ts
@@ -1,7 +1,2 @@
 export * from './useCreateProblemMutation';
 export * from './useEditProblemMutation';
-
-export * from './useContestProblemQuery';
-export * from './useContestProblemSubmissionQuery';
-export * from './useCreateContestProblemSubmissionMutation';
-export * from './useCreateContestProblemSubmissionCheckMutation';
