+++ src/pages/Class/Problem/index.ts
@@ -1,9 +1,5 @@
 export * from './ProblemCreate';
 export * from './ProblemEdit';
 
-export * from './ContestProblemDetail';
-
 export * from './ProblemDescription';
 export * from './ProblemData';
-export * from './ProblemLeaderBoard';
-export * from './ProblemSubmission';
